rewrite implementation
