+++ src/quokka/workflows/tidy.py
@@ -1,13 +1,14 @@
 from __future__ import annotations
 
-from quokka.core.constants import CLANG_TIDY_FILE_EXTENSIONS
 from quokka.core.errors import DiagnosticError
-from quokka.core.types import CommandResult, TidyRequest
+from quokka.core.result import CommandResult
+from quokka.core.types import TidyRequest
+from quokka.model.files import select_tidy_files
 from quokka.model.selectors import resolve_tidy_selector
 from quokka.project.context import CliContext
+from quokka.project.discovery import compile_commands_path
 from quokka.project.state import ensure_no_conflicting_locks
 from quokka.tools.clang_tidy import run_clang_tidy
-from quokka.vcs.git import git_changed_files
 from quokka.workflows.common import summarize_path_examples
 
 
@@ -17,7 +18,7 @@ def run_tidy(context: CliContext, request: TidyRequest) -> CommandResult:
     context.resolve_runtime_dir("tidy")
     ensure_no_conflicting_locks(context, ("build",), "tidy")
 
-    compile_commands = profile.build_dir / "compile_commands.json"
+    compile_commands = compile_commands_path(profile.build_dir)
     if not compile_commands.exists():
         raise DiagnosticError(
             "PROFILE_UNCONFIGURED",
@@ -27,21 +28,19 @@ def run_tidy(context: CliContext, request: TidyRequest) -> CommandResult:
             details={"compile_commands": str(compile_commands)},
         )
 
-    changed_files = git_changed_files(context.worktree_root, selector, "tidy", context.profile_name())
-    files = [path for path in changed_files if path.endswith(CLANG_TIDY_FILE_EXTENSIONS)]
-    skipped_files = [path for path in changed_files if not path.endswith(CLANG_TIDY_FILE_EXTENSIONS)]
-    if not files:
+    selection = select_tidy_files(context, selector)
+    if not selection.files:
         data = {
             "build_dir": str(profile.build_dir),
             "selector": selector,
             "fix": request.fix,
             "files": [],
-            "skipped_files": skipped_files,
+            "skipped_files": selection.skipped_files,
             "no_op": True,
         }
         text = "No files selected for clang-tidy."
-        if skipped_files:
-            text += "\nSkipped non-C/C++ file(s): {}".format(summarize_path_examples(skipped_files))
+        if selection.skipped_files:
+            text += "\nSkipped non-C/C++ file(s): {}".format(summarize_path_examples(selection.skipped_files))
         return CommandResult("tidy", context.profile_name(), None, data, text)
 
     run_clang_tidy(context, selector, fix=request.fix)
@@ -49,10 +48,10 @@ def run_tidy(context: CliContext, request: TidyRequest) -> CommandResult:
         "build_dir": str(profile.build_dir),
         "selector": selector,
         "fix": request.fix,
-        "files": files,
-        "skipped_files": skipped_files,
+        "files": selection.files,
+        "skipped_files": selection.skipped_files,
     }
     text = "Ran clang-tidy wrapper for profile {} with selector '{}'.".format(context.profile_name(), selector)
-    if skipped_files:
-        text += "\nSkipped non-C/C++ file(s): {}".format(summarize_path_examples(skipped_files))
+    if selection.skipped_files:
+        text += "\nSkipped non-C/C++ file(s): {}".format(summarize_path_examples(selection.skipped_files))
     return CommandResult("tidy", context.profile_name(), None, data, text)
